MAE-480: Use only one calculator for fixed membership type
Fixed membership type should always use the same calculator regardless of payment schedule.
