+++ tests/phpunit/CRM/MembershipExtras/Service/MembershipPeriodType/FixedPeriodTypeCalculatorTest.php
@@ -2,16 +2,16 @@
 
 use CRM_MembershipExtras_Service_MembershipTypeDurationCalculator as MembershipTypeDurationCalculator;
 use CRM_MembershipExtras_Service_MembershipTypeDatesCalculator as MembershipTypeDatesCalculator;
-use CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeAnnualCalculator as FixedPeriodCalculator;
+use CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeCalculator as FixedPeriodCalculator;
 use CRM_MembershipExtras_Test_Fabricator_MembershipType as MembershipTypeFabricator;
 use CRM_MembershipExtras_Service_MembershipInstalmentTaxAmountCalculator as MembershipInstalmentTaxAmountCalculator;
 
 /**
- * Class CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeAnnualCalculatorTest
+ * Class CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeCalculatorTest
  *
  * @group headless
  */
-class CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeAnnualCalculatorTest extends BaseHeadlessTest {
+class CRM_MembershipExtras_Service_MembershipPeriodType_FixedPeriodTypeCalculatorTest extends BaseHeadlessTest {
 
   use CRM_MembershipExtras_Test_Helper_FixedPeriodMembershipTypeSettingsTrait;
   use CRM_MembershipExtras_Test_Helper_FinancialAccountTrait;
